+++ client/components/Layout.tsx
@@ -2,25 +2,41 @@ import { ReactNode } from 'react';
 
 import styled from 'styled-components';
 
-import Head from 'next/head';
-
-import RQProvider from '@/services/RQprovider';
-import RenderToast from '@/hooks/toast/RenderToast';
+import RenderToast from '@/components/toast/RenderToast';
+import ReatQueryProvider from '@/services/ReatQueryProvider';
 import NavBar from './Nav/NavBar';
 import StyledProvider from './StyledProvider';
+import PageHead from './PageHead';
+import RenderModal from './modal/RenderModal';
 
-const LayoutStyle = styled.div`
+interface LayoutProps {
+  children: ReactNode;
+}
+
+/** 페이지 레이아웃  */
+function Layout({ children }: LayoutProps) {
+  return (
+    <StyledProvider>
+      <PageHead />
+      <ReatQueryProvider>
+        <LayoutContainer>
+          <NavBar />
+          <Content>{children}</Content>
+        </LayoutContainer>
+      </ReatQueryProvider>
+      <RenderToast />
+      <RenderModal />
+    </StyledProvider>
+  );
+}
+
+export default Layout;
+
+const LayoutContainer = styled.div`
   display: flex;
+  flex-direction: column;
   height: 100vh;
   width: 100vw;
-`;
-
-const Sidebar = styled.div`
-  width: 300px;
-  @media screen and (max-width: 1200px) {
-    width: 100px;
-  }
-  z-index: 1;
 `;
 
 const Content = styled.div`
@@ -28,32 +44,3 @@ const Content = styled.div`
   width: 80%;
   min-width: 0;
 `;
-
-interface LayoutProps {
-  children: ReactNode;
-}
-
-/** 컴포넌트와 navbar의 영역을 분할하고, 전체 컴포넌트의 렌더링을 통제하기 위한 역할 */
-function Layout({ children }: LayoutProps) {
-  return (
-    <StyledProvider>
-      <Head>
-        <title>인자강</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <LayoutStyle>
-        <RQProvider>
-          <Sidebar>
-            <NavBar />
-          </Sidebar>
-          <Content>{children}</Content>
-        </RQProvider>
-      </LayoutStyle>
-      <RenderToast />
-    </StyledProvider>
-  );
-}
-
-export default Layout;
