+++ client/components/PageHead.tsx
@@ -1,0 +1,12 @@
+import Head from 'next/head';
+
+export default function PageHead() {
+  return (
+    <Head>
+      <title>인자강</title>
+      <meta name="description" content="Generated by create next app" />
+      <meta name="viewport" content="width=device-width, initial-scale=1" />
+      <link rel="icon" href="/favicon.ico" />
+    </Head>
+  );
+}
