+++ client/store/toast/index.ts
@@ -1,5 +1,3 @@
 import useToastStore from './useToastStore';
 
-const empty = {};
-
-export { useToastStore, empty };
+export { useToastStore };
