+++ client/types/modal/ModalType.ts
@@ -1,0 +1,11 @@
+type ModalContent = {
+  title: string;
+  message: string;
+};
+
+type ModalAction = (params?: unknown) => void | null;
+
+export type ModalProps = {
+  onAction?: ModalAction;
+  contents: ModalContent;
+};
